remove temporary HitTest module and use common util.
