+++ web/Hoodle/HitTest/Type.hs
@@ -1,0 +1,1 @@
+../../../util/src/Hoodle/HitTest/Type.hs
